requirements.txt
## PDF, OCR, and IMAGE PROCESSING
pdfminer.six>=20221105
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.0

## DATA HANDLING
pandas>=1.5.0
numpy>=1.23.0

## SPELL CHECKING
symspellpy>=6.9.0

## ML AND CLASSIFICATION
scikit-learn>=1.2.0
xgboost>=1.7.0
joblib>=1.2.0

## TEXT MATCHING and PARSING
python-Levenshtein>=0.12.2
regex>=2022.10.31

## DEV and UTILS
tqdm>=4.64.1